api/requirements.txt
Flask
flask-cors
python-dateutil
requests
pandas
sqlalchemy
pytest
